requirements.txt
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-search-results==2.4.2
asyncio==3.4.3
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
yarl==1.9.3
